refactored window into `view.py`
